Added code logic to handle the recyclerview and data binding for each layout. Made also som other changes, see description.
Added glide as dependency for showing images in recyclerview
Added a PhotoGridAdapter for show items in recyclerview
Added BindingAdapters.kt for connecting each item to corresponding layout in binding.
Added code in MainActivity to hide keyboard when edittext is pressed
Added edittext to fragment_overview and made changes for padding and margins.